+++ resource/opentelemetry-resource-detector-azure/CHANGELOG.md
@@ -7,9 +7,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Ensure consistently use of suppress_instrumentation utils
+  ([#2590](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2590))
+
+## Version 0.1.5 (2024-05-16)
+
+- Ignore vm detector if already in other rps
+  ([#2456](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2456))
+- Implement functions resource detector
+  ([#2523](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2523))
+
+## Version 0.1.4 (2024-04-05)
+
+- Fix windows tests/suppress instrumentation for urllib call
+  ([#2178](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2178))
+
+## Version 0.1.3 (2024-01-25)
+
 - Change meta data service timeout to 200ms
   ([#2387](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2387))
 
-## Version 0.1.2 (2024-01-25)
+## Version 0.1.1 (2024-01-10)
 
 - Initial CHANGELOG.md entry